+++ source/items/_item_template.js
@@ -1,0 +1,9 @@
+const { Item } = require("../classes");
+
+const itemName = "";
+module.exports = new Item(itemName, "description",
+	/** specs */
+	(interaction, database) => {
+
+	}
+);